Try to get Keyexchange working
Split verify in verify requests and verify keys. Add a function to
verify keys and get options working for accept, decline the keys or
cancel the verification. if verify keys, verify requests are resetted
and vice versa.
/ircrypt command verify not updated yet.
Also delete the abort function and the ircrypt_requests set, because
there is no way in this programm to send a key without request and if
you have requested a key, its your own fault that you have the work to
decline it.

Signed-off-by: Sven Haardiek <Sven@Haardiek.de>