Handle GPG errors after accepting a key-exchange request
This patch will make IRCrypt check the return code of GPG to decide if
the execution of the GPG command was successful. In case of an error,
the process is aborted, but the request itself is kept in memory so that
it can be used again.

This solves for example the problem of canceling pinentry which would
have caused the key exchange to fail completely before.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>