Issue #20, Print errors to the current buffer
Before, some error messages were printed to the server buffer. For
example if a key to be used for a key exchange did not exist, nothing
happened except for an error being printed to the server buffer where it
would most likely not be noticed.

This commit changes that behavior and prints errors which are a direct
response to a user action into the current buffer (The one which the
command causing the error was entered).

Signed-off-by: Lars Kiesow <lkiesow@uos.de>