Automatically go to IRCrypt buffer
Under some circumstances it is desirable that weechat automatically
switches to the IRCrypt buffer. For example if someone executes the
command “/ircrypt verify requests” it is confusing if nothing happens in
a buffer seen by the user.

This commit lets some commands not only open the buffer but also switch
to the buffer. This can happen every time, or anly if the buffer is
newly opened.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>