Issue #35, Fix deletion of non-existent dictionary entries
If in ircrypt_keyex_get_request or ircrypt_keyex_receive_key the notice
fit in one message, the method would try to delete the buffer with the
previous collected messages which, of couse, does not exist.

Such small messages should not be generated by IRCrypt itself (GPG keys
are just to long), but one could cause trouble by sending a notice to an
IRCrypt user containinge the string “>WCRY-0 ”.

This commit will handle such requests in such a way that IRCrypt will
not interpret them as part of a key exchange anymore, but return them as
just being a normal notice.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>