Added plain command to send unencrypted texts
Sometimes it is desired to send an unencrypted text to a given channel,
even though that one is usually encrypted. Until now, the only way to do
that was to remove the key for this channel and reset it later.

This commit will introduce a “/ircrypt plain ...” command which makes it
easier to send a single unencrypted message to a given server/channel.

Signed-off-by: Sven Haardiek <Sven@Haardiek.de>