Fixed #62, Misuse of encryption marker
Messages are not split properly so that the first part can contain parts
of an unencrypted message.

An attacker without access to the password can use this to create
messages that look like they were encrypted by someone with access to
the password.

Example:

    My message here >CRY-0 some_replayed_message

This commit fixes this problem by ensuring the encryption marker is
placed at the beginning of the message.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>